Moved GOFilename to the loader/GOLoaderFilename
