+++ src/grandorgue/GOOrganController.cpp
@@ -43,6 +43,7 @@
 #include "gui/GOGUIRecorderPanel.h"
 #include "gui/GOGUISequencerPanel.h"
 #include "loader/GOLoadThread.h"
+#include "loader/GOLoaderFilename.h"
 #include "midi/GOMidi.h"
 #include "midi/GOMidiEvent.h"
 #include "midi/GOMidiPlayer.h"
@@ -66,7 +67,6 @@
 #include "GODocument.h"
 #include "GOEvent.h"
 #include "GOFile.h"
-#include "GOFilename.h"
 #include "GOHash.h"
 #include "GOMetronome.h"
 #include "GOOrgan.h"
